requirements.txt
pyheif
pillow
streamlit
